+++ src/main/java/nl/strohalm/cyclos/mfs/models/transactions/TxnLimitResponse.java
@@ -23,9 +23,9 @@ public class TxnLimitResponse implements Serializable {
 
 	private BigDecimal maxAmountPerTxn;
 
-	private Integer maxNumberOfTxnPerDay;
+	private Long maxNumberOfTxnPerDay;
 
-	private Integer maxNumberOfTxnPerMonth;
+	private Long maxNumberOfTxnPerMonth;
 
 	private BigDecimal maxAmountPerDay;
 
@@ -93,19 +93,19 @@ public class TxnLimitResponse implements Serializable {
 		this.maxAmountPerTxn = maxAmountPerTxn;
 	}
 
-	public Integer getMaxNumberOfTxnPerDay() {
+	public Long getMaxNumberOfTxnPerDay() {
 		return maxNumberOfTxnPerDay;
 	}
 
-	public void setMaxNumberOfTxnPerDay(Integer maxNumberOfTxnPerDay) {
+	public void setMaxNumberOfTxnPerDay(Long maxNumberOfTxnPerDay) {
 		this.maxNumberOfTxnPerDay = maxNumberOfTxnPerDay;
 	}
 
-	public Integer getMaxNumberOfTxnPerMonth() {
+	public Long getMaxNumberOfTxnPerMonth() {
 		return maxNumberOfTxnPerMonth;
 	}
 
-	public void setMaxNumberOfTxnPerMonth(Integer maxNumberOfTxnPerMonth) {
+	public void setMaxNumberOfTxnPerMonth(Long maxNumberOfTxnPerMonth) {
 		this.maxNumberOfTxnPerMonth = maxNumberOfTxnPerMonth;
 	}
